# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a chatting API",
        "title": "Jackey Chatting API",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "version": "1"
    },
    "host": "localhost:8888",
    "basePath": "/api/",
    "paths": {
...
        "/friend/delete": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a friend by their user ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Delete a friend",
                "parameters": [
                    {
                        "description": "Friend ID",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/friend.DeleteFriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/friend/friends": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a user's friends",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Get a user's friends",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler._BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.friends"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/friend/manage/addToGroup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Add a friend to a group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend/manage"
                ],
                "summary": "Add a friend to a group",
                "parameters": [
                    {
                        "description": "AddToGroupRequest request",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/friend.AddFriendToGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/friend/manage/createGroup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a friend group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend/manage"
                ],
                "summary": "Create a friend group",
                "parameters": [
                    {
                        "description": "CreateFriendGroupRequest request",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/friend.CreateFriendGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/friend/manage/deleteFromGroup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a friend from a group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend/manage"
                ],
                "summary": "Delete a friend from a group",
                "parameters": [
                    {
                        "description": "DeleteFromGroupRequest request",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/friend.AddFriendToGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/friend/manage/friendGroup": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a user's friend groups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend/manage"
                ],
                "summary": "Get a user's friend groups",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler._BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handler.friendGroup"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/friend/pending": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a user's pending friends",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Get a user's pending friends",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler._BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handler.friends"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Send Pending friend requests",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Friend"
                ],
                "summary": "Send Pending friend requests",
                "parameters": [
                    {
                        "description": "Friend ID",
                        "name": "json_object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/friend.AddFriendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "friend.AddFriendRequest": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "friend.AddFriendToGroupRequest": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "integer"
                },
                "group_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "friend.CreateFriendGroupRequest": {
            "type": "object",
            "properties": {
                "group_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "friend.DeleteFriendRequest": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "handler._BaseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据"
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "handler._ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "err msg"
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "example": "failed"
                }
            }
        },
        "handler.friendGroup": {
            "type": "object",
            "properties": {
                "friends": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.friends"
                    }
                },
                "group_id": {
                    "type": "integer"
                },
                "group_name": {
                    "type": "string"
                }
            }
        },
        "handler.friends": {
            "type": "object",
            "properties": {
                "friend_id": {
                    "type": "integer",
                    "example": 1
                },
                "friend_username": {
                    "type": "string",
                    "example": "user1"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}